+++ lib/paypkg.rb
@@ -49,14 +49,15 @@ private
   # The initialize method reads the config file, calls PayPal for an access_token,
   # stores the access_token in the session, and initializes some variables
   # @param session [Hash] The session object from your ApplicationController subclass
-  def initialize(session=nil)
+  def initialize(session=nil, env=nil, path=nil)
+    env = 'development' if env.nil?
+    path = '.' if path.nil?
     @session = if session then session else {} end
     @session[:paypal_authorization] ||= {}
     @session[:paypal_authorization][:expires_after] ||= Time.now-6
     @session[:paypal_authorization][:access_token] ||= ""
-    env = Rails.env
     @mode = env.to_sym
-    @credentials = YAML.load_file("#{Rails.root.to_path}/config/paypkg.yml")[env]
+    @credentials = YAML.load_file("#{path}/config/paypkg.yml")[env]
     @website = @credentials['website']
     @uri_base = @credentials['uri_base']
     uri = URI.parse(@uri_base)
@@ -87,7 +88,6 @@ private
         @access_token = nil
         raise Net::HTTPServerException.new("Unable to obtain access token from PayPal", response)
       else
-puts response.body.inspect
         hash = JSON.parse(response.body, :symbolize_names=>true)
         @session[:paypal_authorization][:expires_after] = Time.now+hash[:expires_in]
         @session[:paypal_authorization][:access_token] = hash[:access_token]
@@ -152,6 +152,84 @@ public
   # @return [PaypkgResponse Instance]
   def response
     if hash.last then PaypkgResponse.new(hash.last) else nil end
+  end
+
+  # This method collects the response data in such a way as to guarantee
+  # valid Ruby stuctures will be generated, i.e., you won't get a
+  # runtime error because of nil, etc.
+  #
+  # It produces 3 outputs:
+  #  (1) A non-nil error hash
+  #   {
+  #     :name => "VALIDATION_ERROR",
+  #     :details => [
+  #       {
+  #         :field => "payer.funding_instruments[0].credit_card.number",
+  #         :issue => "Value is invalid"
+  #       },
+  #       {
+  #         :field => "payer.funding_instruments[0].credit_card.number",
+  #         :issue => "Number is crap"
+  #       }
+  #     ],
+  #     :message => "Invalid request - see details",
+  #     :information_link => "https://developer.paypal.com/webapps/developer/docs/api/#VALIDATION_ERROR",
+  #     :debug_id => "88dcf0c1730bb"
+  #   }
+  #  (2) A non-nil response object
+  #    <PaypkgResponse:0x00000001e60040 @name="VALIDATION_ERROR",
+  #      @details=[
+  #        <PaypkgResponse:0x00000001e73e10 @field="payer.funding_instruments[0].credit_card.number", @issue="Value is invalid">, \
+  #        <PaypkgResponse:0x00000001e73988 @field="payer.funding_instruments[0].credit_card.number", @issue="Invalid card type"> \
+  #      ],
+  #     @message="Invalid request - see details", \
+  #     @information_link="https://developer.paypal.com/webapps/developer/docs/api/#VALIDATION_ERROR", \
+  #     @debug_id="88dcf0c1730bb"
+  #    >
+  #  (3) A non-nil details array
+  #    [
+  #      {
+  #        :field => "payer.funding_instruments[0].credit_card.number",
+  #        :issue => "Value is invalid"
+  #      },
+  #      {
+  #        :field => "payer.funding_instruments[0].credit_card.number",
+  #        :issue => "Number is crap"
+  #      }
+  #    ]
+  def error_data
+      conditioned_details = []
+      if @hash.last
+        conditioned_hash = @hash.last
+        # sometimes there's details, sometimes not
+        if conditioned_hash.has_key?(:details)
+          # process errors
+          conditioned_hash[:details].each do |detail|
+            conditioned_details << detail
+          end
+        end
+      else
+        status = @status.last
+        case status[0]
+        when '4'
+          name = "CLIENT ERROR"
+          message = "The request sent to PayPal was invalid"
+        when '5'
+          name = "SERVER ERROR"
+          message = "PayPay's servers had a problem"
+        else
+          name = "UNKNOWN ERROR"
+          message = ""
+        end
+        conditioned_hash = {
+            :name => name,
+            :message => "%s. PayPal returned status code %s"%[message,status],
+            :information_link => "https://developer.paypal.com/webapps/developer/docs/api",
+            :debug_id => "0000000000000"
+          }
+      end
+      conditioned_response = PaypkgResponse.new(conditioned_hash)
+      return conditioned_hash, conditioned_response, conditioned_details
   end
 end