+++ test/helpers/init.rb
@@ -1,0 +1,2 @@
+# Require all helpers in the helpers folder
+Dir.glob("#{$app.path}/helpers/*.rb").each { |helper| require("#{helper}") }