+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+0.1.7
+-----
+* Changed the initialize method to pass the database extension (development, test, production), and the database (project) path name. This change was necessary in order to allow this gem to operate without Ruby on Rails.
+
+Before: pp = Paypkg.new(session)
+
+After: pp = Paypkg.new(session, Rails.env, Rails.root.to_path)
+
+0.1.6
+-----
+* Changed the spelling of Polish currency to its US-ASCII equivalent to avoid getting the error 'invalid multibyte char' during require.
+
+0.1.5
+-----
+* Moved the release date into the same file as the version so as not to forget to change it along with the version number.
+
+0.1.4
+-----
+* Added a new method 'error_data' which returns a standardized error response.
+  This method should only be called when 'call_paypal' returns false.
+
+0.1.3
+-----
+* Removed unused parameter from accept_pp_payment.
+* Added code to remove the uri_base from the link provided by PayPal
+  after an authorization in 'validate-credit-card.rb'. Without this change,
+  PayPal will now force close the connection resulting in a ECONNRESET error.
+* Removed an unneeded debugging statement in 'lib/paypkg.rb'.
+
 0.1.2
 -----
 * Exposed the json data (input) string as request, i.e.: